+++ configs/model/mnist.yaml
@@ -1,4 +1,4 @@
-_target_: src.models.mnist.MNISTLitModel
+_target_: src.models.mnist_module.MNISTLitModule
 
 input_size: 784
 lin1_size: 256
