+++ src/datamodules/mnist_datamodule.py
@@ -36,7 +36,6 @@ class MNISTDataModule(LightningDataModule):
         super().__init__()
 
         # this line allows to access init params with 'self.hparams' attribute
-        # it also ensures init params will be stored in ckpt
         self.save_hyperparameters(logger=False)
 
         # data transformations
